e820

+++ include/types.h
@@ -3,12 +3,12 @@
 
 typedef signed char        __s8;
 typedef signed short       __s16;
-typedef signed int         __s32;
+typedef signed long        __s32;
 typedef signed long long   __s64;
 
 typedef unsigned char        __u8;
 typedef unsigned short       __u16;
-typedef unsigned int         __u32;
+typedef unsigned long        __u32;
 typedef unsigned long long   __u64;
 
 #endif	/* _TYPES_H */
